@@ -1,11 +1,16 @@
 package com.hansjoerg.coloringbook.payload;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
-/**
- * Created by rajeevkumarsingh on 02/08/17.
- */
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class LoginRequest {
     @NotBlank
     @Email
@@ -14,19 +19,4 @@ public class LoginRequest {
     @NotBlank
     private String password;
 
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
 }
