@@ -1,16 +1,15 @@
 package com.hansjoerg.coloringbook.security.oauth2.user;
 
+import lombok.Getter;
+
 import java.util.Map;
 
+@Getter
 public abstract class OAuth2UserInfo {
     protected Map<String, Object> attributes;
 
     public OAuth2UserInfo(Map<String, Object> attributes) {
         this.attributes = attributes;
-    }
-
-    public Map<String, Object> getAttributes() {
-        return attributes;
     }
 
     public abstract String getId();
